backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai>=1.66.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
httpx==0.26.0
pydantic==2.6.0
python-multipart==0.0.9
lxml==5.1.0
